week03-homework/pyproject.toml
[project]
name = "week03-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"python-dotenv",
	# add your dependencies here and run `uv sync` to install them
	"llama-index>=0.14.8",
	"llama-index-core>=0.14.8",
	"llama-index-embeddings-dashscope>=0.4.1",
	"llama-index-llms-dashscope>=0.5.1",
	"llama-index-llms-openai-like>=0.5.3",
	"python-dotenv>=1.2.1",
	"paddlepaddle>=3.2.2",
	"paddleocr>=3.3.2",
]
